Refactor autoProcessGames -> games
