+++ nzbToMedia.py
@@ -630,7 +630,7 @@ import sys
 import core
 from core import logger, nzbToMediaDB
 from core.auto_process.comics import Comic
-from core.auto_process.autoProcessGames import Game
+from core.auto_process.games import Game
 from core.auto_process.movies import Movie
 from core.auto_process.music import Music
 from core.auto_process.tv import TV
